requirements.txt
azure-identity==1.25.3
azure-ai-projects==2.0.1
azure-ai-evaluation==1.16.7
azure-mgmt-compute==38.0.0
azure-mgmt-monitor==7.0.0
azure-mgmt-advisor==9.0.0
azure-mgmt-authorization==4.0.0
openai==2.29.0
requests==2.32.5
jupyter==1.1.1
nbconvert==7.17.1
tabulate==0.10.0
pandas==2.3.3,<3.0.0
matplotlib==3.10.9
numpy==2.4.4
pyyaml==6.0.2
python-dotenv==1.1.0
